pyproject.toml
[project]
name = "tg-bot"
version = "0.1.0"
description = "CatSearch bot for the VK hackathon"
authors = [
    {name = "Serge Podkolzin",email = "serzhpodkolzin@ya.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot (>=22.0,<23.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
